删除无效代码，更新版本代码
